+++ analyzer/src/main/java/nl/basjes/parse/useragent/debug/AbstractUserAgentAnalyzerTester.java
@@ -23,7 +23,6 @@ import nl.basjes.parse.useragent.AbstractUserAgentAnalyzer;
 import nl.basjes.parse.useragent.AbstractUserAgentAnalyzerDirect;
 import nl.basjes.parse.useragent.UserAgent;
 import nl.basjes.parse.useragent.UserAgent.MutableUserAgent;
-import nl.basjes.parse.useragent.UserAgentAnalyzerDirect;
 import nl.basjes.parse.useragent.analyze.Matcher;
 import nl.basjes.parse.useragent.analyze.MatchesList.Match;
 import org.slf4j.Logger;
@@ -57,7 +56,7 @@ public class AbstractUserAgentAnalyzerTester extends AbstractUserAgentAnalyzer {
         AbstractUserAgentAnalyzer.configureKryo(kryo);
     }
 
-    public static class KryoSerializer extends UserAgentAnalyzerDirect.KryoSerializer {
+    public static class KryoSerializer extends AbstractUserAgentAnalyzerDirect.KryoSerializer {
         public KryoSerializer(Kryo kryo, Class<?> type) {
             super(kryo, type);
         }
@@ -228,8 +227,8 @@ public class AbstractUserAgentAnalyzerTester extends AbstractUserAgentAnalyzer {
             agent.setUserAgentString(userAgentString);
 
 
-            UserAgent parseResult = null;
-            long      measuredSpeed=-1;
+            UserAgent parseResult   = null;
+            long      measuredSpeed =-1;
             if (measureSpeed) {
                 // Preheat
                 for (int i = 0; i < 100; i++) {